Support bundled ffmpeg libraries in Linux build
ALVR server requires some specific features in libavutil that are not
currently common in usual distributions, mainly vulkan support.

We add a new flag to the build-server xtask: bundle-ffmpeg, this will
download and compile ffmpeg with the required flags.

In server code, a generated wrapper is used to allow loading through
dlopen in case of bundled libraries or standard dynamic linking for the
general case. The link mode is selected at compile time using the
bundled-ffmpeg feature flag.