rework compositor wrapper mechanism
A wrapper is required to inject the ALVR vulkan layer.
For minimal interference, we want the layer to only be present in
vrcompositor process.
We have 2 possible targets: vrcompositor-launcher and vrcompositor, the
first one must have cap_sys_nice=eip, and the second one must retain the
`vrcompositor` name (argv[0]).
Using a shell script is not possible because it can't have the
capability or retain the name when using exec, an ELF is required. It is
better to replace vrcompositor, so there won't be a prompt to fix
capabilities on each run and we can simply rewrite the wrapper on each
execution.