Unbreak Linux audio by selecting 48kHz sampling rate if given choice
Audio servers like PulseAudio that are standard on Linux allow
arbitrary sample rates, all the way down to 1 Hz. The code currently
just chooses the lowest supported sample rate which breaks badly.
Choose 48000 as a reasonable default if given free choice.
(This is the usual native sample rate on most modern systems,
including Linux and Android, and should be a safe choice.)