set b_ref_mode=0 for ffmpeg NVENC encoder
Older NVIDIA cards (e.g. GTX 1650) don't support this feature, and
ffmpeg somehow force-enabled it, despite docs saying it should be set to
"auto", and --help saying the default is "disabled". Setting it to
"disabled" in the encoder works around this issue.

ALVR does not currently use multiple B frames, so there should be no
difference and no obvious difference was visible when testing.

The feature is documented as:
"Using B frame as a reference improves subjective and objective encoded
quality with no performance impact. Hence the users enabling multiple B
frames are strongly recommended to enable this feature."

Co-Authored-By: David Rosca <nowrep@gmail.com>