Update Troubleshooting.md
This fix worked for me. I had two separate computers on the same network with the same issue (unable to detect/ping headset). They worked fine on a separate network. Upon disabling this setting both computers were able to ping the headset and ALVR began to recognize the headset automatically