Clarify link names in template (#2091)
"here" is not a name. It does not describe the link. The main reason this is bad is screen readers often list many/most links in a page and usually aren't going to read out the domain or link. Secondly, you have to read the context around the link to understand what is for. Thirdly, it doesn't say what the link is to or where it's from.